+++ benches/bench-memcmp.rs
@@ -1,6 +1,29 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use std::cmp::Ordering;
 
+// ref.) https://en.wikipedia.org/wiki/Memory_ordering
+fn memory_barrier(_arg: &Vec<&str>) {
+    #[cfg(target_feature = "sse2")]
+    {
+        #[cfg(target_arch = "x86")]
+        use std::arch::x86 as mmx;
+        #[cfg(target_arch = "x86_64")]
+        use std::arch::x86_64 as mmx;
+
+        unsafe { mmx::_mm_mfence() };
+    }
+    /*
+    #[cfg(target_arch = "arm")]
+    {
+        unsafe { core::arch::arm::__dmb(_arg) };
+    }
+    #[cfg(target_arch = "aarch64")]
+    {
+        unsafe { core::arch::aarch64::__dmb(_arg) };
+    }
+    */
+}
+
 #[inline(never)]
 fn process_std_memcmp(texts: &[&str], pattern: &str) -> (usize, usize, usize) {
     let pat_bytes = pattern.as_bytes();
@@ -112,25 +135,30 @@ fn criterion_benchmark(c: &mut Criterion) {
     assert_eq!(n_eq, match_cnt);
     assert_eq!(n_le, less_cnt);
     assert_eq!(n_gr, greater_count);
+    memory_barrier(&vv);
     //
     c.bench_function("std_memcmp", |b| {
         b.iter(|| {
             let _r = process_std_memcmp(black_box(&vv), black_box(pat_string_s));
+            memory_barrier(&vv);
         })
     });
     c.bench_function("memx_memcmp", |b| {
         b.iter(|| {
             let _r = process_memx_memcmp(black_box(&vv), black_box(&pat_string));
+            memory_barrier(&vv);
         })
     });
     c.bench_function("memx_memcmp_basic", |b| {
         b.iter(|| {
             let _r = process_memx_memcmp_basic(black_box(&vv), black_box(&pat_string));
+            memory_barrier(&vv);
         })
     });
     c.bench_function("memx_memcmp_libc", |b| {
         b.iter(|| {
             let _r = process_memx_memcmp_libc(black_box(&vv), black_box(&pat_string));
+            memory_barrier(&vv);
         })
     });
 }